perf: faster builds during development
The plugin starts its processing right before the module optimization phase. This means the
compilation has to wait for the plugin to complete its processing before an app can become usable.
This works fine for the production build but becomes too expensive during development and
hot reloads.

`optimizeModule` hook is currently used to support setups with `ModuleConcatenationPlugin` enabled.
In most setup though, `ModuleConcatenationPlugin` is disabled during development which means we can
move the processing at the end of the compilation process.